# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API DOCUMENTATION",
        "title": "API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "prastiyo.beka12@gmail.com"
        }
    },
    "host": "localhost:3000",
    "paths": {
...
        "/api/v1/company": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Tps",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tps"
                ],
                "summary": "Tps",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Company"
                ],
                "summary": "Company Create",
                "parameters": [
                    {
                        "description": "Company Create",
                        "name": "company",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Company.CompanyRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/issue": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Issue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Issue"
                ],
                "summary": "Issue",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Issue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Issue"
                ],
                "summary": "Issue Create",
                "parameters": [
                    {
                        "description": "Issue Create",
                        "name": "issue",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Issue.IssueRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/issue/delete/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Issue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Issue"
                ],
                "summary": "Issue Delete",
                "parameters": [
                    {
                        "description": "Issue Update",
                        "name": "issue",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Issue.IssueUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Issue ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/issue/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Issue By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Issue"
                ],
                "summary": "Issue By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Issue ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Issue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Issue"
                ],
                "summary": "Issue Update",
                "parameters": [
                    {
                        "description": "Issue Update",
                        "name": "issue",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Issue.IssueUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Issue ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/pemilu": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Pemilu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pemilu"
                ],
                "summary": "Pemilu",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Pemilu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pemilu"
                ],
                "summary": "Pemilu Create",
                "parameters": [
                    {
                        "description": "Pemilu Create",
                        "name": "pemilu",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Pemilu.PemiluRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/pemilu/delete/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Pemilu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pemilu"
                ],
                "summary": "Pemilu Delete",
                "parameters": [
                    {
                        "description": "Pemilu Update",
                        "name": "pemilu",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Pemilu.PemiluUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Pemilu ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/pemilu/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Pemilu By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pemilu"
                ],
                "summary": "Pemilu By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pemilu ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Pemilu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pemilu"
                ],
                "summary": "Pemilu Update",
                "parameters": [
                    {
                        "description": "Pemilu Update",
                        "name": "pemilu",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Pemilu.PemiluUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Pemilu ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/tps": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Tps",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tps"
                ],
                "summary": "Tps Create",
                "parameters": [
                    {
                        "description": "Tps Create",
                        "name": "tps",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Tps.TpsRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/tps/delete/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Tps",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tps"
                ],
                "summary": "Tps Delete",
                "parameters": [
                    {
                        "description": "Tps Update",
                        "name": "tps",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Tps.TpsUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Tps ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/tps/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ambil data Tps By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tps"
                ],
                "summary": "Tps By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tps ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Simpan data Tps",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tps"
                ],
                "summary": "Tps Update",
                "parameters": [
                    {
                        "description": "Tps Update",
                        "name": "tps",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Tps.TpsUpdateRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Tps ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "Company.CompanyRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "dapil_id": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path_photo": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "Issue.IssueRequest": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "issue": {
                    "type": "string"
                },
                "issue_maker": {
                    "type": "string"
                },
                "issue_solution": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "Issue.IssueUpdateRequest": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string"
                },
                "issue": {
                    "type": "string"
                },
                "issue_maker": {
                    "type": "string"
                },
                "issue_solution": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "Pemilu.PemiluRequest": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "info1": {
                    "type": "string"
                },
                "info2": {
                    "type": "string"
                },
                "info3": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "path_photo": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "total_voters": {
                    "type": "integer"
                },
                "total_voters_company": {
                    "type": "integer"
                },
                "total_voters_group": {
                    "type": "integer"
                },
                "total_voters_other": {
                    "type": "integer"
                },
                "tps_id": {
                    "type": "string"
                }
            }
        },
        "Pemilu.PemiluUpdateRequest": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string"
                },
                "info1": {
                    "type": "string"
                },
                "info2": {
                    "type": "string"
                },
                "info3": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "path_photo": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "total_voters": {
                    "type": "integer"
                },
                "total_voters_company": {
                    "type": "integer"
                },
                "total_voters_group": {
                    "type": "integer"
                },
                "total_voters_other": {
                    "type": "integer"
                },
                "tps_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "Tps.TpsRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "dapil_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rw": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "village_id": {
                    "type": "string"
                }
            }
        },
        "Tps.TpsUpdateRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "dapil_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rw": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                },
                "village_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}